requirements.txt
minio==6.0.2
kafka-python==2.0.2
Django==3.1.7